+++ src/components/pages/Comments/Comments.scss
@@ -1,0 +1,8 @@
+.fb-comments{
+  background-color:#e9ebee; 
+}
+
+.comments{
+  color:black;
+  margin-top: 50px;
+}
