+++ src/components/shared/Share/Share.scss
@@ -1,9 +1,9 @@
 
 .fb{
- 
-  border-radius: 55px;
+  border-radius: 32px;
 }
 
 .share{
   color:#4968ad;
+  border-radius: 32px;
 }
